+++ lib/src/log.dart
@@ -1,0 +1,9 @@
+// Copyright (c) 2020, David PHAM-VAN <dev.nfet.net@gmail.com>
+// All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:logging/logging.dart';
+
+/// Private json_intl logger
+final log = Logger('json_intl');